Address CodeRabbit review feedback
- Guard 0-d ndarray in kalman_filter batch_filter (z_i.ndim > 0)
- Cast OpenCV float32 flow points to float64 at MotionEstimator call site
- Widen FixedCamera._background annotation to np.ndarray (dtype from frame)

+++ norfair/camera_motion.py
@@ -589,7 +589,8 @@ class MotionEstimator:
         if curr_pts is not None and prev_pts is not None:
             try:
                 update_prvs, coord_transformations = self.transformations_getter(
-                    curr_pts, prev_pts
+                    curr_pts.astype(np.float64, copy=False),
+                    prev_pts.astype(np.float64, copy=False),
                 )
             except (TypeError, ValueError, np.linalg.LinAlgError) as e:
                 logger.warning(e)
